UserCredential script updated.
Removed temporary scrip. Updated functionality in original script
