+++ Scuti/Scripts/UI/Account/OnboardingView.cs
@@ -13,7 +13,7 @@ namespace Scuti.UI
     {
 
         [SerializeField] Navigation navigation;
-        [SerializeField] UserCreateCredentialForm credentials;
+        [SerializeField] UserCredentialsForm credentials;
         [SerializeField] UserDetailsForm userDetails;
         [SerializeField] AddressForm addressForm;
         [SerializeField] InterestsSelector interestSelect;
