--- Scuti/Scripts/UI/Account/UserCreateCredentialForm.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a6e42cede1b48594eaeab16edbfeb405
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
