+++ Scuti/UI/Prefabs/Account/Onboarding View (Portrait).prefab
@@ -9154,7 +9154,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 0.000000004647344, y: 0}
+  m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0.5}
 --- !u!114 &114101639850121162
@@ -16253,7 +16253,7 @@ GameObject:
   - component: {fileID: 9169347024026273010}
   - component: {fileID: 1889342244225874425}
   m_Layer: 5
-  m_Name: UserCredentialsForm Panel v2
+  m_Name: UserCredentialsForm Panel
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -16304,7 +16304,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2496935428071615601}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a6e42cede1b48594eaeab16edbfeb405, type: 3}
+  m_Script: {fileID: 11500000, guid: f717996bca403384a8e7bf2634ba51f7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   isModal: 0
