+++ Scuti/UI/Prefabs/Account/Onboarding View.prefab
@@ -16634,7 +16634,7 @@ GameObject:
   - component: {fileID: 3379993245127112797}
   - component: {fileID: 1871402104970630095}
   m_Layer: 5
-  m_Name: UserCredentialsForm Panel v2
+  m_Name: UserCredentialsForm Panel
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -16695,7 +16695,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 2489908456630794717}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a6e42cede1b48594eaeab16edbfeb405, type: 3}
+  m_Script: {fileID: 11500000, guid: f717996bca403384a8e7bf2634ba51f7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   isModal: 0
